backend/backend/requirements.txt
asgiref==3.6.0
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
mysql-connector-python==8.0.31
protobuf==3.20.1
pycparser==2.21
PyMySQL==1.0.2
pytz==2022.7
sqlparse==0.4.3
